+++ include/coro_rpc/coro_rpc/common_service.hpp
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include <logging/easylog.h>
+#include <easylog/easylog.h>
 
 #include <filesystem>
