+++ include/coro_rpc/coro_rpc/connection.hpp
@@ -22,7 +22,7 @@
 #include <type_traits>
 #include <utility>
 
-#include "logging/easylog.h"
+#include "easylog/easylog.h"
 #include "rpc_protocol.h"
 #include "util/type_traits.h"
