+++ include/coro_rpc/coro_rpc/coro_connection.hpp
@@ -27,7 +27,7 @@
 #include "asio_util/asio_coro_util.hpp"
 #include "asio_util/asio_util.hpp"
 #include "async_simple/coro/SyncAwait.h"
-#include "logging/easylog.h"
+#include "easylog/easylog.h"
 #include "router.hpp"
 #include "rpc_protocol.h"
 #include "struct_pack/struct_pack.hpp"
