+++ src/coro_rpc/examples/helloworld/server/rpc_service.cpp
@@ -15,7 +15,7 @@
  */
 #include "rpc_service/rpc_service.h"
 
-#include <logging/easylog.h>
+#include <easylog/easylog.h>
 
 #include <thread>
